+++ src/EntityGraphQL/Compiler/EntityQuery/EqlCompileContext.cs
@@ -7,7 +7,14 @@ namespace EntityGraphQL.Compiler.EntityQuery;
 public class EqlCompileContext : CompileContext
 {
     public EqlCompileContext(CompileContext compileContext)
-        : base(compileContext.ExecutionOptions, compileContext.BulkData, compileContext.RequestContext, compileContext.CancellationToken) { }
+        : base(
+            compileContext.ExecutionOptions,
+            compileContext.BulkData,
+            compileContext.RequestContext,
+            compileContext.DocumentVariablesParameter,
+            compileContext.DocumentVariables,
+            compileContext.CancellationToken
+        ) { }
 
     public List<IField> ServiceFieldDependencies { get; } = new();
     public Expression? OriginalContext { get; set; }